% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data_table_with_classes.R
\name{buildAttributes.DataTableWithClasses}
\alias{buildAttributes.DataTableWithClasses}
\title{build the attributes of an object of class DataTableWithClasses based on the count table and pheno table.}
\usage{
\method{buildAttributes}{DataTableWithClasses}(self)
}
\value{
the same object with added fields to describe sample classes and derived
attributes. Sample classes are extracted from the phenoTable, using column(s)
specified in self$classColumn.
}
\description{
Derive the attributes of an object from the count table and pheno table.
This includes the definition of sample and gene names, as well as the association of
classes to samples, based on the classColumn(s) specified in the parameters.
}
\author{
Jacques van Helden and Mustafa AbuElQumsan
}
