% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_attribute.R
\name{buildAttributes.default}
\alias{buildAttributes.default}
\title{default method to build attributes for an object.}
\usage{
\method{buildAttributes}{default}(self)
}
\arguments{
\item{self}{that is an object has been instantiated.}
}
\value{
print message that inform the instantiate the object has been done.
}
\description{
Just send message with object classes. The class-specific builders should have been being called before.
}
\author{
Mustafa AbuElQumsan and Jacques van Helden
}
