% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/count_table_with_classes.R
\name{countTableWithClasses}
\alias{countTableWithClasses}
\title{constructor of the countTableWithClasses class}
\usage{
countTableWithClasses(countTable, phenoTable, dataType, parameters)
}
\arguments{
\item{countTable}{a data.frame with one row per feature (e.g. gene) and
one column per sample.}

\item{phenoTable}{a data frame describing each sample: one row per sample and one column per attribute of a sample.}

\item{parameters}{global and specific parameters for the analysis of this recountID}

\item{dataType="raw}{counts" data type, free text (e.g. raw counts, log2-transformed counts, log2 normalised counts  ...).}
}
\description{
This class contains count tables
(typically used to represent expression profiles in RNA-seq analysis)
documented with a Pheno table (description of each sample) and completed
with some additional fields for convenience: sample names, gene names, ...
}
\author{
Mustafa AbuElQumsan and Jacques van helden
}
