% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data_table_with_classes.R
\name{DataTableWithClasses}
\alias{DataTableWithClasses}
\title{constructor of the DataTableWithClasses class}
\usage{
DataTableWithClasses(dataTable, phenoTable, dataType, parameters)
}
\arguments{
\item{dataTable}{a data.frame with one row per feature (e.g. gene) and
one column per sample.}

\item{phenoTable}{a data frame describing each sample: one row per sample and one column per attribute of a sample.}

\item{parameters}{global and specific parameters for the analysis of this recountID}

\item{dataType="raw}{counts" data type, free text (e.g. raw counts, log2-transformed counts, log2 normalised counts  ...).}
}
\value{
it is create object that is belonge to DataTableWithClasses class that means it would have the follwing attributes
\itemize{
     \item dataTable: is count table that is contains one raw per feature (e.g. gene) and one column per individual (e.g. sample)
     \item phenoTable: that is description table for all individuals whereas that erach raw is individual (sample) and each column is the itemize for such individuals
     \item dataType: that is deemed feature type whereas we have filtered, scaled, log2-norm, PCs of log2norm and so on. to inspect the impact of pre-processing procedures with NGS data.
     \item parameters = that is all the accompanying parameters to recount study case
    \item classColumn: that is indeed refers to type of the outcome column.
    \item nbGenes: that is whole number of features.
    \item sampleNames: that is names of individuals that are included in the new object.
    \item featureNames: the names of each genes in the count table that will be involved in new object.
    \item classLabels: that is the class labels for all the samples in the relatedness object.
    \item classNames:  that is the names of all classes included in the relatedness object.
    \item nbClasses:   that is the overall number of the classes involved in the relatedness object.
    \item classProperities: that is the data.frame composed of two columns one for the class name, number of samples in each class, train size per class
    \item samplesPerClass: such paraperter are involved in classProperities.
    \item classFrequencies: such parameter means the ratio of certain samples for somehow class so as to the whole samples into the dataTable.
    \item randExpectedHitRate: it is corss product for the classFrequensies .
    \item randExpectedMisclassificationRate: it is 1 - randExpectedHitRate.
    \item sampleColors: such parameter for visualisation of the sample by putting some useful colors.
    \item classColors: such parameter for puting special color for each class involved in analysis.
}
}
\description{
This class contains count tables
(typically used to represent expression profiles in RNA-seq analysis)
documented with a Pheno table (description of each sample) and completed
with some additional fields for convenience: sample names, gene names, ...
}
\author{
Mustafa AbuElQumsan and Jacques van helden
}
