% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data_table_with_train_test_sets.R
\name{DataTableWithTrainTestSets}
\alias{DataTableWithTrainTestSets}
\title{Random sampling of n training sets in a DataTableWithClasses.}
\usage{
DataTableWithTrainTestSets(self, stratified = TRUE,
  iterations = parameters$iterations,
  trainingProportion = parameters$trainingProportion)
}
\arguments{
\item{self}{an object of the class DataTableWithClasses}

\item{stratified=TRUE}{if true, sampling is done in each class separately in order to preserve the relative frequencies of classes in training and testing sets.}

\item{iterations=parameters$iterations}{number of  train/test  iterations, which defines the number of independent sampled subsets}

\item{trainingProportion=parameters$trainingProportion}{proportion of samples to sample for each training set}
}
\value{
object of class DataTableWithTrainTestSets containing the following attributes
\itemes {
 \item  iteration: which is number of resampling
  \item trainPropotion:      which is ratio of train volume from whole of dataset
  \item statified:           parameter which refer to stratification criteria would be included in analysis or not.
  \item trainSizePerClass:   which is specify the train volume for each class included in analysis.
  \item trianIndices:        which are whole indices for training process.
  \item testIndices:         which are whole indices for the testing process.

}
}
\description{
Select n random subsets for training. among the biological samples from a DataTableWithClasses.
}
\author{
Jacques van Helden and Mustafa AbuElQumsan
}
