% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/count_table_with_classes.R
\name{defineSampleClasses}
\alias{defineSampleClasses}
\title{define sample classes and all the related attributes (class sizes, nb classes, class colors, class-based sample colors, ...)}
\usage{
defineSampleClasses(self)
}
\value{
the same object with added fields to describe sample classes and derived
attributes. Sample classes are extracted from the phenoTable, using column(s)
specified in self$classColumn.
}
\author{
Jacques van Helden and Mustafa AbuElQumsan
}
