% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ErrorRateBoxPlot.R
\name{ErrorRateBoxPlot}
\alias{ErrorRateBoxPlot}
\title{Function for boxplot the error rate}
\usage{
ErrorRateBoxPlot(experimentList, classifier,
  main = paste("Misclassification rates", "\\n", parameters$iterations,
  "iterations"), expColor = "#00BBFF", permColor = "grey",
  expLegend = "Train/test", permLegend = "Permuted labels", data.type,
  variable.type = "all.variables", ....)
}
\arguments{
\item{experimentList}{this is list of experiment where each cell of list error rate of the single experiment with special parameters}

\item{classifier}{such are represent which classifier have been used in the analysis}

\item{main}{is the main title of the boxplot}

\item{expColor}{is colore each box plot that represent the error rate for each experiment with actual class lables.}

\item{permColor}{is the coler for the box plot that represent error rate for each experiment with permuted calss lables.}

\item{expLegend}{is the legend for the real class lable train/test experiment.}

\item{permLegend}{legend for the one experiment with permuted class lables.}

\item{data.type}{which is any data type used for testing the error rate form the specified classifier}

\item{variable.type}{this indicate for the knid and amount of the viariable that are used in analysis}
}
\description{
this function takes the experiment List from the error rate that are collected from the experiment.
}
\examples{
compareExperiments(experimentList = train.test.results)
}
\author{
Mustafa AbuElqumsan and Jacques van Helden
}
