% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/count_table_with_classes.R
\name{exportTables.countTableWithClasses}
\alias{exportTables.countTableWithClasses}
\title{Export the different fields of an object of class countTableWithClasses in tab-separated values text files.}
\usage{
\method{exportTables}{countTableWithClasses}(self, export.dir, file.prefix,
  extension = ".tsv")
}
\arguments{
\item{self}{an object, which must belong to a compatible class}

\item{export.dir}{export directory}

\item{file.prefix}{file prefix to build the different tables}

\item{extension=".tsv"}{extension for the expodrted files (tab-separated value)}

\item{export.dir}{export directory}

\item{file.prefix}{file prefix to build the different tables}

\item{extension=".tsv"}{extension for the expodrted files (tab-separated value)}
}
\author{
Mustafa AbuElQumsan and Jacques van Helden
}
