% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exportTables.R
\name{exportTables.countTableWithTrainTestSets}
\alias{exportTables.countTableWithTrainTestSets}
\title{Export the tables specific to the class countTableWithTrainTestSets}
\usage{
\method{exportTables}{countTableWithTrainTestSets}(self, export.dir,
  file.prefix, extension = ".tsv")
}
\arguments{
\item{self}{an object, which must belong to class countTableWithTrainTestSets}

\item{export.dir}{export directory}

\item{file.prefix}{file prefix to build the different tables}

\item{extension=".tsv"}{extension for the exported files (tab-separated values)}
}
\author{
Jacques van Helden
}
