% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export_tables.R
\name{exportTables.DataTableWithClasses}
\alias{exportTables.DataTableWithClasses}
\title{Export the different fields of an object of class DataTableWithClasses in tab-separated values text files.}
\usage{
\method{exportTables}{DataTableWithClasses}(self,
  export.dir = self$parameters$dir$tsv, file.prefix, extension = ".tsv")
}
\arguments{
\item{self}{an object, which must belong to a compatible class}

\item{export.dir}{export directory}

\item{file.prefix}{file prefix to build the different tables}

\item{extension=".tsv"}{extension for the exported files (tab-separated values)}
}
\author{
Mustafa AbuElQumsan and Jacques van Helden
}
