% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export_tables.R
\name{exportTables.DataTableWithTrainTestSets}
\alias{exportTables.DataTableWithTrainTestSets}
\title{Export the tables specific to the class DataTableWithTrainTestSets}
\usage{
\method{exportTables}{DataTableWithTrainTestSets}(self, export.dir, file.prefix, extension = ".tsv")
}
\arguments{
\item{self}{an object, which must belong to class DataTableWithTrainTestSets}

\item{export.dir}{export directory}

\item{file.prefix}{file prefix to build the different tables}

\item{extension=".tsv"}{extension for the exported files (tab-separated values)}
}
\description{
Export the tables specific to the class DataTableWithTrainTestSets
}
\author{
Mustafa AbuElQumsan and Jacques van Helden.
}
