% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export_tables.R
\name{exportTables.default}
\alias{exportTables.default}
\title{default export table (for closing the NextMethod calls)}
\usage{
\method{exportTables}{default}(self, ...)
}
\arguments{
\item{self}{an object, which must belong to a compatible class}
}
\description{
to print message which is inform us the exporting processes has been done.
}
\author{
Jacques van Helden and Mustafa AbuELQumsan.
}
