% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/count_table_with_classes.R
\name{exportTables}
\alias{exportTables}
\title{Export the different fields of an aobject in tab-separated values text files.}
\usage{
exportTables(self, ...)
}
\arguments{
\item{self}{an object, which must belong to a compatible class}
}
\author{
Mustafa AbuElQumsan and Jacques van Helden
}
