% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_pheno_table.R
\name{ExtractPhenoTable}
\alias{ExtractPhenoTable}
\title{Fuction for extracting Pheno Table from specified recount project}
\usage{
ExtractPhenoTable(rse, verbose = FALSE)
}
\arguments{
\item{rse}{this is Range summaraised Experiment object which is in the Rdata file}

\item{verbose}{for testing}
}
\description{
Once we have the pheno table from assigned project through recount data set.
}
\author{
Mustafa ABUELQUMSAN and Jacques van Helden
}
