% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file_prefix.R
\name{filePrefix}
\alias{filePrefix}
\title{file prefix procedures with a given classifier and a given object}
\usage{
filePrefix(self, classifier, permute = FALSE, file.prefix = NULL)
}
\arguments{
\item{self}{an object of class DataTableWithTrainTestSets}

\item{classifier}{is the type of classifier that is used with repeated process.}

\item{permute}{is show if the class lable are permuted this for sake of the knowing the strength and weaknesses of the classifier}

\item{file.prefix}{in order to let us to save file for any other script}
}
\description{
due to such file will be called from other scripts in order to computes some classifier-specific parameters
}
\author{
Mustafa ABUELQUMSAN and Jacques van Helden
}
