% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filterCountTable.R
\name{filterCountTable}
\alias{filterCountTable}
\title{Apply various filters on the observations and variables of
a count table in order to prepare it for classification.}
\usage{
filterCountTable(countTable, phenoTable, classLabels,
  minSamplesPerClass = parameters$minSamplesPerClass,
  nearZeroVarFilter = parameters$nearZeroVarFilter, draw.plot = TRUE)
}
\arguments{
\item{countTable}{a data.frame containing the number of read counts per gene (row) in each sample (column)}

\item{phenoTable}{a data.frame describing each sample (row) with different attributes (columns)}

\item{classLabels}{a vector indicating the class associated to each sample.}

\item{minSamplesPerClass=parameters$minSamplesPerClass}{minimum nuimber of samples per class to keep}

\item{nearZeroVarFilter=parameters$nearZeroVarFilter}{if TRUE, applyb caret::nearZeroVariance() to filter out poor predictor genes}

\item{draw.plot=TRUE}{if TRUE, draw an histogram of variance per gene.}
}
\description{
Apply various filters on the observations (e.g. biological samples)
and variables (e.g. genes) of
a count table in order to prepare it for classification.
}
\examples{
## Load a data set
countdata <- loadRecountExperiment(recountID = "SRP048759", mergeRuns = TRUE)
countTable <- t(countdata$merged$sampleCounts)
phenoTable <- countdata$merged$samplePheno

# Run the filtering
filteredData <- filterCountTable(countTable, phenoTable, classColumn="tissue")

# Replace unfiltered data by filtered data
countTable <- filteredData$countTable
phenoTable <- filteredData$phenoTable
classLabels <- filteredData$classLabels

## Filter a dataset and build classLabels based on 2 columns
countdata <- loadRecountExperiment(recountID = "SRP057196", mergeRuns = TRUE)
countTable <- t(countdata$merged$sampleCounts)
phenoTable <- countdata$merged$samplePheno

#' # Run the filtering
filteredData <- filterCountTable(countTable, phenoTable, classColumn=c("tissue", "cell.type"), minSamplesPerClass=5)
table(filteredData$classLabels)

}
\author{
Mustafa AbuElQumsan and Jacques van Helden
}
