% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_data_table.R
\name{filterDataTable}
\alias{filterDataTable}
\title{Apply various filters on the observations and variables of
a count table in order to prepare it for classification.}
\usage{
filterDataTable(rawCounts, draw.plot = TRUE)
}
\arguments{
\item{countsWithClasses}{an object of the class DataTableWithClasses.
This object contains the data table + the parameters (including filtering parameters).}

\item{draw.plot=TRUE}{if TRUE, draw an histogram of variance per gene.}
}
\description{
Apply various filters on the observations (e.g. biological samples)
and variables (e.g. genes) of
a count table in order to prepare it for classification.
}
\examples{
## Load a data set
countdata <- loadRecountExperiment(recountID = "SRP048759", mergeRuns = TRUE)
dataTable <- t(countdata$merged$sampleCounts)
phenoTable <- countdata$merged$samplePheno

# Run the filtering
filteredData <- filterDataTable(dataTable, phenoTable, classColumn="tissue")

# Replace unfiltered data by filtered data
dataTable <- filteredData$dataTable
phenoTable <- filteredData$phenoTable
classLabels <- filteredData$classLabels

## Filter a dataset and build classLabels based on 2 columns
countdata <- loadRecountExperiment(recountID = "SRP057196", mergeRuns = TRUE)
dataTable <- t(countdata$merged$sampleCounts)
phenoTable <- countdata$merged$samplePheno

#' # Run the filtering
filteredData <- filterDataTable(dataTable, phenoTable, classColumn=c("tissue", "cell.type"), minSamplesPerClass=5)
table(filteredData$classLabels)

}
\author{
Mustafa AbuElQumsan and Jacques van Helden
}
