% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_recount_experiment.R
\name{geocharFromPheno}
\alias{geocharFromPheno}
\title{Extract the conditions from the "characteristics" column of the coldata.}
\usage{
geocharFromPheno(runPheno)
}
\description{
This is a bit tricky: we have to parse a string describing several attributes.
}
