% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/train_test_result.R
\name{heatmap.TrainTestResult}
\alias{heatmap.TrainTestResult}
\title{heat map is a fales color image with pattern added to the left side and to the top.}
\usage{
heatmap.TrainTestResult(object)
}
\arguments{
\item{object}{is the object belonge to TrainTestresult that have summary about the results of Training-Testing evaluation of the given classifier.}
}
\value{
heatmap figure for the object belonge to the TrainTestresult.
}
\description{
to show reordering of the rows and columns according to some set of values(row or column means) within the restrictions imposed by pattern is carried out.
}
\author{
Jacques van Helden and Mustafa AbuElQumsan
}
