% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/study_case.R
\name{histCountsPerGeneClass}
\alias{histCountsPerGeneClass}
\title{Draw an histogram with the raw counts per gene}
\usage{
histCountsPerGeneClass(self)
}
\arguments{
\item{self}{object belong to StudyCase class.}
}
\value{
an object of the same class as the input object
}
\description{
to carify the distribution of the raw counts per gene.
}
\author{
Mustafa AbuElQumsan and Jacques van Helden
}
