% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/study_case.R
\name{histCountsPerGeneClass.StudyCase}
\alias{histCountsPerGeneClass.StudyCase}
\title{Draw an histogram with the raw  counts per gene}
\usage{
\method{histCountsPerGeneClass}{StudyCase}(self)
}
\arguments{
\item{self}{object belong to StudyCase class.}
}
\value{
an object of the same class as the input object
}
\author{
Jacques van Helden & Mustafa AbuElQumsan
}
