% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init_parallel_computing.R
\name{initParallelComputing}
\alias{initParallelComputing}
\title{initialise parameters for parallel computing with foreach and doParallel packages}
\usage{
initParallelComputing(project.parameters)
}
