% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init_recount_id.R
\name{initRecountID}
\alias{initRecountID}
\title{Intialise parameters and directories for a give recount ID.}
\usage{
initRecountID(recountID, project.parameters)
}
\arguments{
\item{recountID}{must be a valid ID row Recount2 database (SRP...)}

\item{project.parameters}{project parameters (read from the YAML-formatted file)}
}
\value{
all project.parameters for the related recount IDn as specified in the YAML configuration file.
}
\description{
This function ensures the consistency of the parameters for a giveb RNA-seq experiment specified by its recount ID.
(1) Merge the global parameters with experiment-specific parameters
(description, short label, ...).
(2) Check that some required parameters are well defined.
(3) Check the output directory (if not specified in config, define them with default paths) and create them if required.
}
\author{
Jacques van Helden and Mustafa AbuElQumsan
}
