% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iterate_training_testing.R
\name{IterateTrainingTesting.DataTableWithTrainTestSets}
\alias{IterateTrainingTesting.DataTableWithTrainTestSets}
\title{Iterate training/testing procedures with a given classifier and a given data type, with object has class DataTableWithTrainTestSets.}
\usage{
\method{IterateTrainingTesting}{DataTableWithTrainTestSets}(dataset,
  classifier, permute = FALSE, file.prefix = paste(sep = "_",
  dataset$ID, dataset$dataType, classifier), verbose = 1)
}
\arguments{
\item{dataset}{an object of class DataTableWithTrainTestSets}

\item{classifier}{is the type of classifier that is used with repeated process.}

\item{permute}{is show if the class lable are permuted this for sake of the knowing the strength and weaknesses of the classifier}

\item{file.prefix}{prefix for files. If NULL, files will not be saved.}

\item{verbose=1}{level of verbosity}
}
\value{
an object which is Misclassification error rate for the specified number for resampling
    \itemize{
        \item testTable: that is the table that is contains misclassification error rate for specified  number of resampling. for an object belonge to DataTableWithTrainTestSets.
      }
}
\description{
for sake of the accuracy and due to the error rate have computed from sampleing from the origin count data,
# it is better to compute the the error rate multiple time and then we would find the avarage for the error rate. with DataTableWithTrainTestSets's object
}
\author{
Mustafa AbuElQumsan and Jacques van Helden
}
