% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iterate_training_testing.R
\name{IterateTrainingTesting}
\alias{IterateTrainingTesting}
\title{Iterate training/testing procedures with a given classifier and a given data type}
\usage{
IterateTrainingTesting(dataset, ...)
}
\arguments{
\item{dataset}{an object of class DataTableWithTrainTestSets}

\item{classifier}{is the type of classifier that is used with repeated process.}

\item{permute}{is show if the class lable are permuted this for sake of the knowing the strength and weaknesses of the classifier}

\item{file.prefix}{in order to let us to save file from the IterateTrainingTesting}
}
\description{
for sake of the accuracy and due to the error rate have computed from sampleing from the origin count data,
# it is better to compute the the error rate multiple time and then we would find the avarage for the error rate.
}
\author{
Mustafa ABUELQUMSAN and Jacques van Helden
}
