% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/libsize_barplot.R
\name{LibsizeBarplot}
\alias{LibsizeBarplot}
\title{Draw a barplot with the number of million reads per sample}
\usage{
LibsizeBarplot(counts, sample.labels = colnames(counts),
  sample.colors = NULL, main = "Library sizes",
  xlab = "libsum (Million reads per sample)", cex.names = 1,
  lmargin = cex.names * max(nchar(sample.labels))/2.5 + 5, ...)
}
\arguments{
\item{counts}{a count table with 1 row per feature and 1 column per sample}

\item{...}{additional parameters are passed to barplot}

\item{sample.labels=colnames(counts)}{sample labels (displayed as y legend)}

\item{sample.colors=NULL}{sample-specific colors (e.g. reflecting conditions)}

\item{main="Library}{sizes" Main title for the plot}

\item{xlab="libsum}{(Million reads per sample)" X axis label}

\item{plot.file=NULL}{Path of a file to store the figure (pdf format).}

\item{fig.format="pdf"}{File format (only used if plto.file is defined)}

\item{cex.names=1}{fondt size for sample labels}

\item{lmargin=cex.names*max(nchar(sample.labels))/2.5+5}{Adapt margin to the label sizes}

\item{width=7}{width of the plot file}

\item{height=length(sample.labels)/3+2}{height of the plot file (if not specified, will be adapted to the number of samples)}
}
\description{
Draw a barplot with the number of million reads per sample
}
\author{
Jacques van Helden
}
