% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/libsize_rank_plot.R
\name{LibsizeRankPlot}
\alias{LibsizeRankPlot}
\title{Rank plot of RNA-seq library sizes}
\usage{
LibsizeRankPlot(
  count.table,
  parameters,
  plot.file = NULL,
  plot.width = 7,
  plot.heigh = 5
)
}
\arguments{
\item{count.table}{a table of counts per feature, with one row per feature (e.g. gene, transcript) and one column per library (e.g. run, sample).}

\item{plot.file}{if ot null, save th plot in pdf format in the specified file}

\item{plot.width=7}{width of the pdf file (only works when plot.file is not null)}

\item{plot.height=5}{height of the pdf file (only works when plot.file is not null)}
}
\description{
Plot the sizes (in million reads) of RNA-seq librairies ranked by decreasing size.
}
\author{
Jacques van Helden
}
