% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_counts.R
\name{loadCounts}
\alias{loadCounts}
\title{Loading the count table, the Pheno table and vector of classes}
\usage{
loadCounts(recountID = parameters$recountID,
  classColumn = parameters$classColumn, mergeRuns = parameters$mergeRuns,
  sampleIdColumn = parameters$sampleIdColumn,
  minSamplesPerClass = parameters$minSamplesPerClass,
  dir.workspace = parameters$dir$workspace, ...)
}
\arguments{
\item{recountID}{number of experiment}

\item{classColumn}{## Column(s) of the pheno table used to define class labels. Passed to filterCountTable()}

\item{mergeRuns}{to wrapper the rus by sample}

\item{...}{additional parameters are passed to loadRecountExperiment()}

\item{sampleIdColumn="geo_accession"}{name of the column of the pheno table which contains the sample IDs.}

\item{minSamplesPerClass=parameters$minSamplesPerClass}{min number of samples per class for the filtering.  Passed to filterCountTable()}

\item{dir.workspace=parameters$dir$workspace}{Directory in which the data and results will be stored.}
}
\value{

}
\description{
Given a ReCount ID, load the count table, the pheno table
and the vector of classes, prepares and filters the count table, via
wrapper runs by sample.
}
\examples{

##############################################
## Load a RNA-seq dataset and merge counts per sample
x <- loadCounts( recountID = "SRP048759", mergeRuns=T, classColumn = "tissue")

################################################################
## RNA-seq data combining 2 columns to define sample classes
## Reduce the min number of samples per class for this dataset.
x <- loadCounts( recountID = "SRP057196", mergeRuns=T,
    classColumn = c("tissue", "cell.type"), minSamplesPerClass=5)

}
\author{
Mustafa ABUELQUMSAN and Jacques van Helden
}
