% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_recount_experiment.R
\name{loadRecountExperiment}
\alias{loadRecountExperiment}
\title{Load one count table from ReCount and optionally merge the counts per sample.}
\usage{
loadRecountExperiment(recountID = parameters$recountID,
  dir.workspace = parameters$dir$workspace,
  mergeRuns = parameters$mergeRuns,
  sampleIdColumn = parameters$sampleIdColumn,
  classColumn = parameters$classColumn, forceDownload = FALSE,
  verbose = parameters$verbose, ...)
}
\arguments{
\item{recountID=parameters$recountID}{identifier of one study in ReCount database}

\item{dir.workspace=parameters$dir$workspace}{path of the folder to store the results (in one separate sub-directory per recountID)}

\item{mergeRuns=parameters$mergeRuns}{if TRUE, read counts will be merged for each sample}

\item{sampleIdColumn=parameters$sampleIdColumn}{name of the column of the pheno table which contains the sample IDs.
This information is passed to MergeRuns().}

\item{verbose=TRUE}{if TRUE, write messages to indicate the progressing of the tasks}

\item{forceDownload=FALSE}{by default, the data is downloaded only if it is not found in the studyPath folder.
If forceDownload is TRUE, the data will be downloaded irrespective of existing files.}
}
\value{
A list containing: the count table, the pheno table, and some additional parameters (study ID, ...).
}
\description{
Load one count table from Recount for a given experiment ID, and optionally
merge the counts in order to avoid redundancy between multipl runs per sample.
}
\examples{

## Load counts for a Leukemia dataset
recountID <- "SRP048759"
recountData <- loadRecountExperiment(recountID)

## Check the dimension of the table with counts per run
dim(recountData$result$runs$countTable)

## Check the number of runs per sample
table(recountData$runPheno$geo_accession)

## Count the number of unique sample IDs in the run-wise description table (runPheno)
message("Number of runs = ", length(recountData$runPheno$geo_accession))
message("Number unique geo_accession values: ", length(unique(sort(recountData$runPheno$geo_accession))))

## Check the dimension of the table with counts per sample
dim(recountData$merged$sampleCounts)

cor(log(recountData$result$runs$countTable[, recountData$runPheno$geo_accession == "GSM1521620"]+1))

## Test correlation between 8 randomly selected biological replicates (distinct samples)
cor(log(recountData$merged$sampleCounts[, sample(1:ncol(recountData$merged$sampleCounts), size=8)]+1))


}
\author{
: Jacques van Helden and Mustafa Abuelqumsan
}
