% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_required_bioconductor_packages.R
\name{LoadRequiredBioconductorPackages}
\alias{LoadRequiredBioconductorPackages}
\title{Load required packages from BioConductor, afte having installed them if necessary.}
\usage{
LoadRequiredBioconductorPackages(packages, verbose = 0)
}
\arguments{
\item{packages}{a vector containing a list of required BioConductor packages.}
}
\value{
chack and then install all BioConductor packages if it aren't installed.
}
\description{
Load a list of required BioConductor packages. For each specified package,
first check  if it is available on this R instance. If not, automatically install it.
}
\author{
Mustafa AbuElQumsan and Jacques van Helden
}
