% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_required_cran_packages.R
\name{LoadRequiredCRANPackages}
\alias{LoadRequiredCRANPackages}
\title{Load required packages from CRAN, after having installed them if necessary.}
\usage{
LoadRequiredCRANPackages(packages, verbose = 0)
}
\arguments{
\item{packages}{a vector containing a list of required CRAN packages.}
}
\value{
chack and then install all CRAN packages if it aren't installed.
}
\description{
Load a list of required RCRAN packages. For each specified package,
first check  if it is available on this R instance. If not, automatically install it.
}
\author{
Mustafa AbuElQumsan and Jacques van Helden
}
