% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merge_runs.R
\name{MergeRuns}
\alias{MergeRuns}
\title{Merge read counts from different runs by sample.}
\usage{
MergeRuns(countsPerRun, phenoTable, sampleIdColumn = "geo_accession",
  verbose = FALSE)
}
\arguments{
\item{countsPerRun}{a data frame containing read counts, one column per run and one row per gene}

\item{pheno.table}{a data frame containing the description of each run. This must include a column with the sample ID associated to each run}

\item{verbose=FALSE}{if TRUE, write messages to indicate the progressing of the tasks}
}
\description{
In ReCount data, some samples are sequenced over several runs, which increases the coverage but creates problems
# because technical replicates come from the same sample and are thus not independent.
# We thus merge the runs per sample, in order to obtain a single vector of read counts per sample.
}
\author{
Jacques van Helden
}
