% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/message_with_time.R
\name{message.with.time}
\alias{message.with.time}
\title{print a message preceded by currrent tume}
\usage{
message.with.time(...)
}
\arguments{
\item{...}{all parameters are passed to the message() function}
}
\author{
Jacques van Helden
}
