% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalize_count_table.R
\name{NormalizeCounts}
\alias{NormalizeCounts}
\title{Normalize RNA-seq count table}
\title
\usage{
NormalizeCounts(rawCounts, method = "quantile", quantile = 0.75,
  log2 = TRUE, epsilon = 0.1)
}
\arguments{
\item{rawCounts}{a raw count table, with one column per sample and one row per gene.
Normalization will be done by samples (columns).}

\item{method="quantile"}{normalization method. Supported: sum, mean, media, quantile.
Preferred method: quantile = 0.75. Quantiles are more robust to outliers than the mean
or sum. The median is sometimes weak in RNA-seq data, the 75th percentile (quantile 0.75)
seems a good tradeoff between robustness and representativity (taking into account a
representative proportion of the total counts per samples).
Note that median is the same as quantile 0.5.}

\item{quantile=0.75}{quantile used as scaling factor when quantile method is selected.}

\item{log2=TRUE}{apply a log2 transformation}

\item{epsilon=0.1}{value added to all counts before applying the log2 transformation
in order to avoid zero counts.}
}
\value{
a list with the normalized count table + a table providing sample-wise statistics
}
\description{
Normalize a table of raw counts from RNA-seq
}
\examples{

}
\author{
Jacques van Helden and Mustafa AbuElqumsan
}
