% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/output_parameters.R
\name{outputParameters}
\alias{outputParameters}
\title{Compute output parameters for an RNAseqMVA analysis}
\usage{
outputParameters(dataset, classifier, permute = FALSE,
  createDir = TRUE)
}
\arguments{
\item{dataset}{an object of class DataTableWithTrainTestSets}

\item{classifier}{is the type of classifier that is used with repeated process.}

\item{permute=FALSE}{is show if the class lable are permuted this for sake of the knowing the strength and weaknesses of the classifier}

\item{createDir=TRUE}{if TRUE, the result directory is created automatically if it does not exist}
}
\value{
a list with the following fields
\itemize{
\item resultDir: path to the result directory, in which the result files should be stored.
\item filePrefix: string that should prepend the name of each result file.
\item figureLabel:  short string with a label for the graphics
}
}
\description{
Compute output parameters (output directory, file prefix, figure label)  for a given analysis given its input parameterrs (recountIR, dataset, classifier, permutation or not).
}
\author{
Mustafa AbuElQumsan and Jacques van Helden
}
