% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_2pcs.R
\name{plot2PCs}
\alias{plot2PCs}
\title{Plot two components of prcomp transform}
\usage{
plot2PCs(self, main = paste0(self$ID, " ", self$parameters$feature, " ",
  self$parameters$short_label, "\\nPCs of ", self$dataType),
  col = self$sampleColors, pcs = c(1, 2), ...)
}
\arguments{
\item{self}{an object of class DataTableWithClasses containing a field prcomp (produced by prcompt function)}

\item{main}{main title is built by default by can be replaced}

\item{col}{= self$sampleColors dots are colored according to sample class membership.}

\item{...}{additional parameters are passed to plot()}

\item{pcs=c(1, 2)}{the two indices of the principal components to be displayed}
}
\description{
Plot the first and second components of a prcomp() result in a lighter way than the default biplot.
}
\author{
Jacques van Helden and Mustafa AbuELQumsan
}
