% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotfigures.R
\name{plotFigures.countTableWithClasses}
\alias{plotFigures.countTableWithClasses}
\title{plot diagnostic figures for an object of the class CountTableWithClasses.}
\usage{
\method{plotFigures}{countTableWithClasses}(self, plot.dir, file.prefix,
  dataType = self[["dataType"]], extension = ".pdf")
}
