% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotfigures.R
\name{plotFigures.DataTableWithClasses}
\alias{plotFigures.DataTableWithClasses}
\title{plot diagnostic figures for an object of the class DataTableWithClasses.}
\usage{
\method{plotFigures}{DataTableWithClasses}(self, plot.dir, file.prefix,
  dataType = self[["dataType"]], extension = ".pdf")
}
