% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotfigures.R
\name{plotFigures}
\alias{plotFigures}
\title{plot diagnostic figures for specific classes of objects}
\usage{
plotFigures(self, ...)
}
\arguments{
\item{self}{is an object that is belonge to DataTableWithClasses.}
}
\value{
utmost figures and visualisation about the main statistics and the distribution for the raw data.
}
\description{
to explore the nature of raw data and showing the the distribution and main statistics for the raw data.
}
\author{
Mustafa AbuElQUmsan and Jacques van Helden
}
