% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_filter_histogram.R
\name{plotFilterHistograms}
\alias{plotFilterHistograms}
\title{plot histograms of the variance distributions at different steps of the filtering process}
\usage{
plotFilterHistograms(dataset, rawCounts, plot.file = NULL,
  plot.height = NULL)
}
\arguments{
\item{dataset}{an object of class DatasetWithClasses returned by filterDataTable()}

\item{plot.file=NULL}{save the plot in a specified pdf file}
}
\description{
plot histograms of the variance distributions at different steps of the filtering process
}
\author{
Jacques van Helden
}
