% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_pc12.R
\name{plotPC12}
\alias{plotPC12}
\title{Plot first versus second components of prcomp transform}
\usage{
plotPC12(self, main = paste(sep = "", self$ID, "; PCs of log2norm counts"),
  col = self$sampleColors, ...)
}
\arguments{
\item{self}{an object of class countTableWithClasses containing a field prcomp (produced by prcompt function)}

\item{main}{main title is built by default by can be replaced}

\item{col}{= self$sampleColors dots are colored according to sample class membership.}

\item{...}{additional parameters are passed to plot()}
}
\description{
Plot the first and second components of a prcomp() result in a lighter way than the default biplot.
}
\author{
Jacques van Helden
}
