% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tune_classifiers.R
\name{plotRFtuning}
\alias{plotRFtuning}
\title{draw plots to highlight the results of ranfomForest tuning}
\usage{
plotRFtuning(tuningResult, pdfFile = NULL, ...)
}
\arguments{
\item{tuningResult}{result of TuneClassifiers() run with option tuneRandomForest=TRUE, and where vectors of values were provided for mtry and ntree}

\item{...}{additional parameters are passed to plot()}

\item{pdfFile=NULL}{if not null, the plot is exported in pdf format the specified file. Else it is printed out on the current device.}
}
\value{
no return field
}
\description{
draw a series of plots highlighting the impact of the parameters (mtry, ntree) on the performances (miclassification error rate) of randomForest
}
\author{
Jacques van Helden
}
