% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tune_classifiers.R
\name{plotSVMtuning}
\alias{plotSVMtuning}
\title{draw plots to highlight the results of SVM tuning}
\usage{
plotSVMtuning(tuningResult, pdfFile = NULL, ...)
}
\arguments{
\item{tuningResult}{result of TuneClassifiers() run with option tuneSVM=TRUE, and where vectors of values were provided for the different parameters}

\item{...}{additional parameters are passed to plot()}

\item{pdfFile=NULL}{if not null, the plot is exported in pdf format the specified file. Else it is printed out on the current device.}
}
\value{
no return field
}
\description{
draw a series of plots highlighting the impact of the parameters on the performances (miclassification error rate) of SVM
}
\author{
Jacques van Helden
}
