% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data_table_with_classes.R
\name{print.DataTableWithClasses}
\alias{print.DataTableWithClasses}
\title{finish summary of an object belonging to class DataTableWithClasses}
\usage{
\method{print}{DataTableWithClasses}(x)
}
\arguments{
\item{self}{that is an object which is want to print its properities}
}
\value{
properities of an object
}
\description{
this is default generic funtion in order to print the properities of the object  that is belonge to DataTableWithClasses.
}
\author{
Mustafa AubElQumsan and Jacques van Helden
}
