% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/train_test_result.R
\name{print.TrainTestResult}
\alias{print.TrainTestResult}
\title{printing the attribute of object that had TraiTestResult class.}
\usage{
\method{print}{TrainTestResult}(object)
}
\arguments{
\item{object}{is the object belonge to TrainTestresult that have summary about the results of Training-Testing evaluation of the given classifier.}
}
\value{
showcase some the object belonge to the TrainTestresult that are Test (misclassifiaction rate) and Train (learning error).
}
\description{
showcasting the some useful attributes of the object had TrainTestResult in order to seen the results of Training-Testing evaluation of the given classifier.
}
\author{
Mustafa AbuElQumsan and Jacques van Helden.
}
