% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/required_libraries.R
\name{RequiredCRANPackages}
\alias{RequiredCRANPackages}
\title{Load required packages from CRAN, and install them if necessary.}
\usage{
RequiredCRANPackages(packages)
}
\arguments{
\item{packages}{a vector containing a list of required CRAN packages.}
}
\description{
Load a list of required RCRAN packages. For each specified package, 
first check  if it is available on this R instance. If not, automatically install it.
}
\author{
Mustafa AbuElQumsan and Jacques van Helden
}
