% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/study_case.R
\name{RunDESeq2}
\alias{RunDESeq2}
\title{run DESeq2 to test differential expression on each feature of a data table.}
\usage{
RunDESeq2(self)
}
\arguments{
\item{self}{object}
}
\value{
an object of the same class as the input object
}
