% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/study_case.R
\name{RunedgeR}
\alias{RunedgeR}
\title{run edgeR to test differential expression on each feature of a data table.}
\usage{
RunedgeR(self)
}
\arguments{
\item{self}{object}
}
\value{
an object of the same class as the input object
}
