% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/study_case.R
\name{RunViRf}
\alias{RunViRf}
\title{Compute variable importance with Random Forest on an object of class StudyCase, and}
\usage{
RunViRf(self)
}
\arguments{
\item{self}{an object belonging to StudyCase class.}
}
\value{
a clone of the input StudyCase object with an added
DataTableWithTrainTestSets containing the log2-normalized counts data table
where features have been re-ordered by decreasing the importance of the features.
}
\description{
run variable importance by random Forest to test importance on each feature of a data table.
}
\author{
Mustafa AbuElQumsan and Jacques van Helden
}
