% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/study_case.R
\name{RunViRf.StudyCase}
\alias{RunViRf.StudyCase}
\title{run variable importance by random Forest on an object of class StudyCase}
\usage{
\method{RunViRf}{StudyCase}(self)
}
\arguments{
\item{self}{object belong to StudyCase class.}
}
\value{
a clone of the input StudyCase object with an added
DataTableWithTrainTestSets containing the log2norm data table
where features have been re-ordered by decreasing the importance of the features.
}
\description{
run variable importance by random Forest on an object of class StudyCase to test importance on each feature of a data table, and order variables by decreasing the importance of features in dataTable.
}
\author{
Mustafa AbuElQumsan and Jacques van Helden
}
