% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/memory_images.R
\name{saveMemoryImage}
\alias{saveMemoryImage}
\title{Save a memory image in order to be able to reload all results without having to recompute everything}
\usage{
saveMemoryImage(memory.dir = parameters$dir$memoryImages,
  recountID = parameters$recountID, prefix = "RNAseqMVA_image_")
}
\author{
Jacques van Helden anf Mustafa AbuElQumsan
}
