% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select_projects_by_characteristics.R
\name{selectProjectsByCharacteristics}
\alias{selectProjectsByCharacteristics}
\title{Select Recount projects based on the presence of user-specified fields in their characteristics.}
\usage{
selectProjectsByCharacteristics(recount.metadata, query.charact,
  run.kable = FALSE, run.barplot = FALSE)
}
\arguments{
\item{recount.metadata}{a data.frame containing the metadata, as returned by recount::all_metadata() function.}

\item{query.charact}{a string indicating the query to be matched against field names in the characteristics tables.}

\item{run.kable=FALSE}{if TRUE, print out the summary of the matches in kable format for markdown report.}

\item{run.barplot=FALSE}{if TRUE, draw a barplot with the number of runs per selected project}
}
\description{
Recount metadata contains a field named "characteristics", which is a list of fields with values.
It is useful to run a query on the field names (not values) in order to select projects
related to a given experimental setting (e.g. "time series"), or to a particular field (e.g. "disease type").
}
\examples{
## Load recount metadata
library(recount)
recount.metadata <- all_metadata(subset='sra')

## Select all the projects havinf "disease status" as a characteristics field
selectProjectsByCharacteristics(recount.metadata,
   query.charact = "disease status",
   run.barplot = TRUE)
}
\author{
Mustafa AbuElQumsan and Jacques van Helden
}
