% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/count_table_with_classes.R
\name{selectStratifiedTrainingSets}
\alias{selectStratifiedTrainingSets}
\title{Stratified random sampling of n training sets among samples.}
\usage{
selectStratifiedTrainingSets(countTable, iterations = parameters$iterations,
  trainingProportion = parameters$trainingProportion)
}
\arguments{
\item{self}{an object of the class CountTableWithClasses}

\item{iterations=parameters$iterations}{number of  train/test  iterations, which defines the number of independent sampled subsets}

\item{trainingProportion=parameters$trainingProportion}{proportion of samples to sample for each training set}
}
\description{
Select n random subsets of the samples for training,  distributed across the classes according to a user-specified training proportion.
}
\author{
Jacques van Helden and Mustafa AbuElQumsan
}
