% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/count_table_with_classes.R
\name{selectTrainingSets.countTableWithClasses}
\alias{selectTrainingSets.countTableWithClasses}
\title{Random sampling of n training sets in a CountTableWithClasses.}
\usage{
\method{selectTrainingSets}{countTableWithClasses}(self, stratified = TRUE,
  iterations = parameters$iterations,
  trainingProportion = parameters$trainingProportion)
}
\arguments{
\item{self}{an object of the class CountTableWithClasses}

\item{stratified=TRUE}{if true, sampling is done in each class separately in order to preserve the relative frequencies of classes in training and testing sets.}

\item{iterations=parameters$iterations}{number of  train/test  iterations, which defines the number of independent sampled subsets}

\item{trainingProportion=parameters$trainingProportion}{proportion of samples to sample for each training set}
}
\description{
Select n random subsets for training. among the biological samples from a CountTableWithClasses.
}
\author{
Jacques van Helden and Mustafa AbuElQumsan
}
