% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/count_table_with_classes.R
\name{selectTrainingSets}
\alias{selectTrainingSets}
\title{Export the object with train/test sets by stratification sampling}
\usage{
selectTrainingSets(self, ...)
}
\arguments{
\item{self}{which much belong to the countTableWithClasses class.}
}
\description{
sampling is done in each class separately in order to preserve the relative frequencies of classes in training and testing sets
}
\author{
Mustafa ABUELQUMSAN and Jacques van Helden
}
