% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stratified_sampling.R
\name{StratifiedSampling}
\alias{StratifiedSampling}
\title{stratified sampling}
\usage{
StratifiedSampling(labels, size = length(labels))
}
\arguments{
\item{labels}{a vector with labels to be resampled}

\item{size=length(labels)}{size of the output (resampled) vector}
}
\description{
Perform a stratified sampling from a vector of group labels:
each initial group is replaced by random labels proportional to the initial
group proportions.
}
\examples{

## Generate a vector of group labels
labels <- rep(x = c("A", "B", "C", "D"), times = c(256, 64, 16, 4))

## Non-stratified sampling
nstr <- sample(labels, size = length(labels) , replace = FALSE)
table(labels, nstr)
## Repeat this several times: the proportions vary between iterations

## Stratified permutation of group labels
str1 <- StratifiedSampling(labels)
## str1 has same proportions as labels (except for rounding of group sizes)
table(labels, str1)

## Second stratified permutation
str2 <- StratifiedSampling(labels)
## str2 has same proportions as labels (except for rounding of group sizes)
table(labels, str2)

## str1 and str2 are different and independent
table(str1, str2)

## Subsampling
str3 <- StratifiedSampling(labels, size = length(labels) * 2/3)
## str3 has same proportions as labels (except for rounding of group sizes)
table(str3)

}
\author{
Jacques.van-Helden@univ-amu.fr
}
