% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/study_case.R
\name{StudyCase}
\alias{StudyCase}
\title{Create a StudyCase object, load RecountID dataset, and generate datasets.}
\usage{
StudyCase(recountID, parameters)
}
\arguments{
\item{recountID}{A valide ID for an object of the ReCount2 database}

\item{parameters}{recountID-specific parameters specified in a YAML-formatted configuration file}
}
\value{
an object of class StudyCase
}
\description{
Create a StudyCase object, load RecountID dataset, and run preprocessing to generate the DataTableWitthClasses objects.
}
\details{
#' @title build DataTableWithTrainTestSets for an instance StudyCase depending on its class
#' @author Mustafa AbuElqumsan and Jacques van Helden
#' @export

DataTableWithTrainTestSets.StudyCase <- function(dataSet){
  message("\tInstantiate 'DataTableWithTrainTestSets' for object of class ", paste(collapse=", ", class(dataSet)))
  UseMethod("DataTableWithTrainTestSets", dataSet)
  return(dataSet)

}
}
\author{
Jacques van Helden and Mustafa AbuElQumsan
}
