% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize_error_tables.R
\name{SummarizeErrorTable}
\alias{SummarizeErrorTable}
\title{generate a summary table from a list of error table files.}
\usage{
SummarizeErrorTable(files, stopIfMissing = TRUE, labels = files)
}
\arguments{
\item{files}{a list of files with the error rates for one classification analysis.}

\item{stopIfMissing=TRUE}{if TRUE, issue an error if any of the input files does not exist}

\item{labels=files}{labels associated to each file (used for displays)
The files should be formatted as tab-separated values.
Each row corresponds to one iteration.}
}
\description{
read a list of error tables and summarize their results in
a single summary table with 1 row per input file and one column per statistics
of the original error tables. The summary statistics is the mean, except for
the iteration, where the max value of the error table is taken.
}
\author{
Jacques van Helden and Mustaga AbuElQumsan
}
