% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize_train_test_results.R
\name{SummarizeTrainTestResults}
\alias{SummarizeTrainTestResults}
\title{Summarize results of several training-testing experiments.}
\usage{
SummarizeTrainTestResults(experimentList,
  experimentLabels = names(experimentList), verbose = 1)
}
\arguments{
\item{experimentList}{a list of results returned by IterateTrainingTesting()}

\item{verbose=1}{level of verbosity}
}
\description{
Extract the results of a list containing training-testing
results produced by the function IterateTrainingTesting().
}
\examples{

recountID <- "SRP042620"
summaryTable <- SummarizeTrainTestResults(
  experimentList = train.test.results.knn.k.values[[recountID]])

}
\author{
Jacques.van-Helden@univ-amu.fr
}
