% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/count_table_with_classes.R
\name{summary.countTableWithClasses}
\alias{summary.countTableWithClasses}
\title{print a summary of an object belonging to class countTableWithClasses}
\usage{
summary.countTableWithClasses(x)
}
\author{
Mustafa AubElQumsan and Jacques van Helden
}
