% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/count_table_with_train_test_sets.R
\name{summary.countTableWithTrainTestSets}
\alias{summary.countTableWithTrainTestSets}
\title{summary for a countTableWithTrainTestSets object}
\usage{
summary.countTableWithTrainTestSets(x)
}
