% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/count_table_with_classes.R
\name{summary.DataTableWithClasses}
\alias{summary.DataTableWithClasses}
\title{print a summary of an object belonging to class DataTableWithClasses}
\usage{
summary.DataTableWithClasses(self)
}
\author{
Mustafa AubElQumsan and Jacques van Helden
}
