% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/count_table_with_train_test_sets.R
\name{summary.DataTableWithTrainTestSets}
\alias{summary.DataTableWithTrainTestSets}
\title{summary for a DataTableWithTrainTestSets object}
\usage{
summary.DataTableWithTrainTestSets(x)
}
