% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/study_case.R
\name{summary.StudyCase}
\alias{summary.StudyCase}
\title{print a summary of an object belonging to class StudyCase}
\usage{
summary.StudyCase(self)
}
\value{
print the summary of such object by utilizing generic function
}
\description{
just print the summary of the object that is belonge to class StudyCase
}
\author{
Mustafa AubElQumsan and Jacques van Helden
}
