% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/study_case.R
\name{XYplot}
\alias{XYplot}
\title{Draw an XY plot to compare the mean counts per gene for two user-specified class}
\usage{
XYplot(self)
}
\arguments{
\item{self}{object belong to StudyCase class.}
}
\value{
an object of the same class as the input object
}
\description{
Drawing an XY plot to compare the mean counts per gene for two user-specified class of the raw counts per gene.
}
\author{
Mustafa AbuElQumsan and Jacques van Helden
}
