% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/study_case.R
\name{XYplot.StudyCase}
\alias{XYplot.StudyCase}
\title{Draw an XY plot to compare the mean counts per gene for two user-specified class}
\usage{
\method{XYplot}{StudyCase}(self, class1, class2)
}
\arguments{
\item{self}{object belong to StudyCase class.}

\item{class1}{the name of the first class to plot}

\item{class2}{the name of the second class to plot}
}
\value{
an object of the same class as the input object
}
\description{
Drawing an XY plot to compare the mean counts per gene for two user-specified class of the raw counts per gene.
}
\author{
Mustafa AbuElQumsan and Jacques van Helden
}
