% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_table_with_train_test_sets.R
\name{buildAttributes.DataTableWithTrainTestSets}
\alias{buildAttributes.DataTableWithTrainTestSets}
\title{build the attributes that depend on the count table}
\usage{
\method{buildAttributes}{DataTableWithTrainTestSets}(self)
}
\description{
#################### Builde constractor for the selecting Training Sets #########
#' @title Export the object with train/test sets by stratification sampling
#' @author Mustafa ABUELQUMSAN and Jacques van Helden
#' @description sampling is done in each class separately in order to preserve the relative frequencies of classes in training and testing sets
#' @param self which much belong to the DataTableWithClasses class.
#'
#' @export
selectTrainingSets <- function(self, ...){
  message("\tExporting the object class  ", class(self), " object with train/test sets")
  UseMethod("selectTrainingSets", self)
}
}
