% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_attribute.R
\name{buildAttributes.default}
\alias{buildAttributes.default}
\title{default method to build attributes for an object.}
\usage{
buildAttributes.default(self)
}
\description{
Just send message with object classes. The class-specific builders should have been be called before.
}
