% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/count_table_with_classes.R
\name{buildAttributes.default}
\alias{buildAttributes.default}
\title{default method to build attributes for an object.}
\usage{
\method{buildAttributes}{default}(self)
}
\description{
Just send message with object classes. The class-specific builders should have been be called before.
}
