% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deg_ordering.R
\name{DEGordering}
\alias{DEGordering}
\title{Variable ordering according to the p-values returned by differential expression analysis.}
\usage{
DEGordering(dataTableWithClasses, method = "DESeq2", randomized = TRUE,
  edgeRDispEst = "tagwise")
}
\arguments{
\item{dataTableWithClasses}{an object containing both the data table and the class label vector.}

\item{method="DESeq2"}{choice of method for differential expression analysis. Supported: "DESeq2" , "edgeR".}

\item{edgeRDispEst="tagwise"}{method used by edgeR to estimate the dispersion. Supported: common, trended, tagwise.}

\item{...}{all additional parmeters are passed to the differentail expression method (DEseq2 or edgeR).}
}
\value{
a list with 2 tables:
1) result$geneOrder: a vector of gene names ordered by increasing p-value.
2) result$orderedDataTable ordered count table where columns (genes) have been ordered by increasing p-value.
3) result$DEGtable a table with one row per gene, and one column per DEG statistics (mean, log-ratio, nominal and adjusted p-values, ...)

Note that genes are ordered according to nominal p-value (pvalue) rather than
adjusted p-value (padj) because DESeq2 produces NA values for the padj.
}
\description{
Order the variables (columns) of a count table in preparation for supervised
classification, by running differential expression analysis with either DESeq2 or edgeR.
}
\details{
###################################################
}
\examples{

###################################################
## loading required packages
LoadRequiredBioconductorPackages(packages = c("DESeq2","edgeR"))

recountID <- "SRP048759"
studyCases <- loadCounts(recountID = recountID, mergeRuns = T, classColumn = "tissue")
degOrderdPValues <- DEGordering(dataTableWithClasses = studyCases$dataTable, classes =  studyCases$classes, method = "edgeR")
## degPValues<- degPValues[order(degPValues$padj) ,]

choose the method of differential expression analysis

}
\author{
Mustafa AbuElQumsan and Jacques van Helden
}
