% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/error_rate_box_plot.R
\name{ErrorRateBoxPlot}
\alias{ErrorRateBoxPlot}
\title{Function for boxplot the error rate}
\usage{
ErrorRateBoxPlot(experimentList, classifier, data.type, fig.height = 10,
  fig.width = 4 + 0.2 * length(experimentList),
  experimentLabels = names(experimentList), main = paste(sep = "",
  parameters$recountID, "; ", classifier, "\\n", data.type, "; ",
  parameters$iterations, "iterations"), boxplotFile = NULL,
  expColor = "#00BBFF", permColor = "grey", expLegend = "Train/test",
  permLegend = "Permuted labels", ....)
}
\arguments{
\item{experimentList}{this is list of experiment where each cell of list error rate of the single experiment with special parameters}

\item{classifier}{such are represent which classifier have been used in the analysis}

\item{main}{is the main title of the boxplot}

\item{expColor}{is colore each box plot that represent the error rate for each experiment with actual class lables.}

\item{permColor}{is the coler for the box plot that represent error rate for each experiment with permuted calss lables.}

\item{expLegend}{is the legend for the real class lable train/test experiment.}

\item{permLegend}{legend for the IterateTrainingTesting with permuted class lables.}

\item{dataType}{data type(s), used to build the file prefix}

\item{variablesType}{variable types (e.g. all, DEG, ...), used to build the file prefix}
}
\description{
this function takes the experiment List from the error rate that are collected from the experiment.
}
\examples{
compareExperiments(experimentList = train.test.results)
}
\author{
Mustafa AbuElqumsan and Jacques van Helden
}
