% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/error_rate_box_plot.R
\name{ErrorRateBoxPlot}
\alias{ErrorRateBoxPlot}
\title{Function for boxplot the error rate}
\usage{
ErrorRateBoxPlot(experimentList, classifier,
  expMisclassificationRate = NULL, horizontal = TRUE,
  fig.height = 10, fig.width = 4 + 0.05 * length(experimentList),
  cex.axis = 0.8, experimentLabels = names(experimentList),
  main = paste0(parameters$short_label, " (", parameters$recountID, ") ",
  parameters$feature, "s", "\\n", classifier, "; ", parameters$iterations,
  " iterations"), boxplotFile = NULL, medianLine = FALSE,
  expColor = "#00BBFF", permColor = "grey", expLegend = "Train/test",
  permLegend = "Permuted labels", legend.place = "topright", ....)
}
\arguments{
\item{experimentList}{this is list of experiment where each cell of list error rate of the single experiment with special parameters}

\item{classifier}{name of the classifier (e.g. svm, knn, rf), used in the default main title}

\item{expMisclassificationRate}{user-specified value for the expected misclassification rate}

\item{main}{main title for the boxplot}

\item{expColor}{is colore each box plot that represent the error rate for each experiment with actual class lables.}

\item{permColor}{is the coler for the box plot that represent error rate for each experiment with permuted calss lables.}

\item{expLegend}{is the legend for the real class label train/test experiment.}

\item{permLegend}{legend for the IterateTrainingTesting with permuted class lables.}

\item{legend.place}{location for the legend, which is passed to legend()}

\item{medianLine=FALSE}{if TRUE, draw a line joining the medians}
}
\description{
this function takes the experiment List from the error rate that are collected from the experiment.
}
\examples{
compareExperiments(experimentList = train.test.results)
}
\author{
Mustafa AbuElqumsan and Jacques van Helden
}
