% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export_tables.R
\name{exportTables}
\alias{exportTables}
\title{Export the different fields of an object in tab-separated values text files.}
\usage{
exportTables(self, ...)
}
\arguments{
\item{self}{an object, which must belong to a compatible class}
}
\description{
Such target for ease the function for the biologists by converting the Rdata object for the tab-separated delimited file for facilitate the difficulties for extrating dataTables and phenoTables
}
\author{
Mustafa AbuElQumsan and Jacques van Helden
}
