% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export_tables.R
\name{exportTables.StudyCase}
\alias{exportTables.StudyCase}
\title{Export the tables specific to the class StudyCase}
\usage{
\method{exportTables}{StudyCase}(self, export.dir = self$parameters$dir$tsv,
  extension = ".tsv")
}
\arguments{
\item{self}{an object, which must belong to class StudyCase}

\item{export.dir}{export directory}

\item{file.prefix}{file prefix to build the different tables}

\item{extension=".tsv"}{extension for the exported files (tab-separated values)}
}
\description{
this calls exportTable on each of the datasets  (which belong to DataTableWithClasses and DataTableWithTrainTestSets)
}
\author{
Jacques van Helden
}
