% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filterCountTable.R
\name{filterCountTable}
\alias{filterCountTable}
\title{Apply various filters on the observations and variables of
a count table in order to prepare it for classification.}
\title
\usage{
filterCountTable(countTable, phenoTable, classColumn = "tissue",
  minSamplesPerClass = 10, ...)
}
\arguments{
\item{countTable}{a data.frame with one row per observation and one column per variable}

\item{phenoTable}{a data.frame with one row per observation and one column per attribute}

\item{classColumn="tissue"}{name of a column of the pheno table which contains the class labels (default: "tissue").
In som cases classes must be built by concatenating several columns of the pheno table (e.g. "tissue" and "cell.type" for dataset SRP057196),
This can be achieved by providing a vector of column names from the pheno table. In this case, class names are built by}

\item{minSamplesPerClass=10}{minimum nuimber of samples per class to keep}
}
\description{
Apply various filters on the observations (e.g. biological samples)
and variables (e.g. genes) of
a count table in order to prepare it for classification.
}
\examples{
## Load a data set
countdata <- loadRecountExperiment(recountID = "SRP048759", mergeRuns = TRUE)
countTable <- t(countdata$merged$sampleCounts)
phenoTable <- countdata$merged$samplePheno

# Run the filtering
filteredData <- filterCountTable(countTable, phenoTable, classColumn="tissue")

# Replace unfiltered data by filtered data
countTable <- filteredData$countTable
phenoTable <- filteredData$phenoTable
classes <- filteredData$classes

## Filter a dataset and build classes based on 2 columns
countdata <- loadRecountExperiment(recountID = "SRP057196", mergeRuns = TRUE)
countTable <- t(countdata$merged$sampleCounts)
phenoTable <- countdata$merged$samplePheno

#' # Run the filtering
filteredData <- filterCountTable(countTable, phenoTable, classColumn=c("tissue", "cell.type"), minSamplesPerClass=5)
table(filteredData$classes)

}
\author{
Mustafa AbuElQumsan and Jacques van Helden
}
