% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geochar_from_pheno.R
\name{geocharFromPheno}
\alias{geocharFromPheno}
\title{Extract the conditions from the "characteristics" column of Recount pheno Table}
\usage{
geocharFromPheno(runPheno)
}
\arguments{
\item{runPheno}{it is the Run pheno table which may be containt duplicate in the rus for the same sample, and it give us the actual length for a pheno table}
}
\value{
phenoTable with data.frame properities in order to be much earier in handle it smoothly.
}
\description{
This is a bit tricky: we have to parse a string describing several attributes.
}
\author{
Mustafa ABUELQUMSAN and Jacques van Helden
}
