% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geocharFromPheno.R
\name{geocharFromPheno}
\alias{geocharFromPheno}
\title{Extract the conditions from the "characteristics" column of the coldata.}
\usage{
geocharFromPheno(runPheno)
}
\arguments{
\item{runPheno}{it is the Run pheno table which may be containt duplicate in the rus for the same sample, and it give us the actual length for a pheno table}
}
\description{
This is a bit tricky: we have to parse a string describing several attributes.
}
\author{
Mustafa ABUELQUMSAN and Jacques van Helden
}
