% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geochar_from_pheno.R
\name{geocharFromPheno}
\alias{geocharFromPheno}
\title{Extract the conditions from the "characteristics" column of Recount pheno Table, and store them in separate columns.}
\usage{
geocharFromPheno(runPheno)
}
\arguments{
\item{runPheno}{run pheno table produced by colData(), which contains the pheno information for each run (there might be several runs per )}
}
\value{
phenoTable with data.frame properities in order to be much earier in handle it smoothly.
}
\description{
This is a bit tricky: we have to parse a string describing several attributes.
}
\author{
Mustafa AbuElQumsan and Jacques van Helden
}
