% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iterate_training_testing.R
\name{IterateTrainingTesting}
\alias{IterateTrainingTesting}
\title{Iterate training/testing procedures with a given classifier and a given data type}
\usage{
IterateTrainingTesting(dataset, ...)
}
\arguments{
\item{dataset}{an object of class DataTableWithTrainTestSets}

\item{classifier}{is the type of classifier that is used with repeated process.}

\item{permute}{if TRUE, permute the calss labels to measure misclassifciation rate without relevant training}

\item{file.prefix}{in order to let us to save file from the IterateTrainingTesting}
}
\value{
an object which is Misclassification error rate for the specified number for resampling
\itemize{
     \item testTable: that is the table that is contains misclassification error rate for specified  number of resampling.

}
}
\description{
for sake of the accuracy and due to the error rate have computed from sampleing from the origin count data,
# it is better to compute the the error rate multiple time and then we would find the avarage for the error rate.
}
\author{
Mustafa AbuElQumsan and Jacques van Helden
}
