% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_counts.R
\name{loadCounts}
\alias{loadCounts}
\title{Loading the count table, the Pheno table and vector of class labels}
\usage{
loadCounts(recountID, parameters, ...)
}
\arguments{
\item{recountID}{number of experiment}

\item{...}{additional parameters are passed to loadRecountExperiment()}

\item{paramters}{list of parameters (supposedly loaded from the yaml configuration file)}
}
\value{
it is returen the two feature types:
 1) for the original counts that means implicitly the raw count table before making ang pre-processing proccess.
 2) for the filtered count that means implicitly the raw count after made the pre-processing proccess,

  \itemize {
     \item filtered: count table after applying the pre-processing proccess.

  }
}
\description{
Given a ReCount ID, load the count table, the pheno table
and the vector of class labels, prepares and filters the count table, via
wrapper runs by sample.
}
\examples{

##############################################
## Load a RNA-seq dataset and merge counts per sample
x <- loadCounts( recountID = "SRP048759", mergeRuns=T, classColumn = "tissue")

################################################################
## RNA-seq data combining 2 columns to define sample classes (classLabels vector)
## Reduce the min number of samples per class for this dataset.
x <- loadCounts( recountID = "SRP057196", mergeRuns=T,
    classColumn = c("tissue", "cell.type"), minSamplesPerClass=5)

## RCurl and XML are Required for recount but not declared in dependencies
}
\author{
Mustafa ABUELQUMSAN and Jacques van Helden
}
