% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_counts.R
\name{loadCounts}
\alias{loadCounts}
\title{Loading the count table, the Pheno table and vector of class labels}
\usage{
loadCounts(recountID, parameters, ...)
}
\arguments{
\item{recountID}{number of experiment}

\item{parameters}{list of parameters (supposedly loaded from the yaml configuration file)}

\item{...}{additional parameters are passed to loadRecountExperiment()}
}
\value{
it is returen the two feature types:
\enumerate{
\item for the original counts that means implicitly the raw count table before making ang pre-processing proccess.
\item for the filtered count that means implicitly the raw count after made the pre-processing proccess,
}

\itemize{
\item filtered: count table after applying the pre-processing proccess.

}
}
\description{
Given a ReCount ID, load the count table, the pheno table
and the vector of class labels, prepares and filters the count table, via
wrapper runs by sample.
}
\examples{

## Load an RNA-seq dataset and merge counts per sample.
# This assumes that the required parameters have been specified
# in the YAML configuration file.

library("RNAseqMVA")
source('misc/01a_load_libraries.R')
source('misc/01b_load_parameters.R')
recountID <- selectedRecountIDs[1] ## We only load one recountID at a time
parameters <- initRecountID(recountID, project.parameters)
x <- loadCounts( recountID = "SRP048759", parameters = parameters)

## RCurl and XML are Required for recount but not declared in dependencies
}
\author{
Mustafa ABUELQUMSAN and Jacques van Helden
}
