% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/memory_images.R
\name{loadMemoryImage}
\alias{loadMemoryImage}
\title{Load a previous analysis previously stored as memoy image}
\usage{
loadMemoryImage(memory.dir = parameters$dir$memoryImages,
  recountID = parameters$recountID, prefix = "RNAseqMVA_image_")
}
\author{
Jacques van Helden anf Mustafa AbuElQumsan
}
