% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merge_runs.R
\name{MergeRuns}
\alias{MergeRuns}
\title{Merge read counts from different runs by sample.}
\usage{
MergeRuns(runs)
}
\arguments{
\item{runs}{an object of the class DataTableWithClasses}

\item{sampleIdColumn="geo_accession"}{name of the column of the pheno table which contains the sample IDs. GEO accession is preferred because it is widely used, but it is sometimes not defined. In such case, the "sample" column should be used.}

\item{verbose=FALSE}{if TRUE, write messages to indicate the progressing of the tasks}
}
\value{
this is original cout table witout multipl runs per sample
 \itemize{
   \item mergedRuns that is the count table after marging the runs to be original count table after get ride of the multipl run per sample.
 }
mergedRuns
}
\description{
In ReCount data, some samples are sequenced over several runs, which increases the coverage but creates problems
 # because technical replicates come from the same sample and are thus not independent.
# We thus merge the runs per sample, in order to obtain a single vector of read counts per sample.
}
\author{
Jacques van Helden and Mustafa AbuELQumsan
}
