% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/message_with_time.R
\name{message.with.time}
\alias{message.with.time}
\title{print a message preceded by currrent tume}
\usage{
message.with.time(...)
}
\arguments{
\item{...}{all parameters are passed to the message() function}
}
\value{
report message with time for each proccess.
}
\description{
such function in order to be able to compute the time has been exploit through the computation process.
}
\author{
Jacques van Helden and Mustafa AbuElQumsan
}
