% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/message_with_time.R
\name{message.with.time}
\alias{message.with.time}
\title{print a message preceded by currrent time}
\usage{
message.with.time(...)
}
\arguments{
\item{...}{all parameters are passed to the message() function}
}
\value{
report message with time for each proccess.
}
\description{
Print a message prepended with the current time in reverse order (YYYY-MM-DD) in order to match alphabetical order with chronological order.
}
\author{
Jacques van Helden and Mustafa AbuElQumsan
}
