% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misclassification_estimate.R
\name{MisclassificationEstimate}
\alias{MisclassificationEstimate}
\title{RNA-Seq classifier evaluation to assesse the performance of the classifier}
\usage{
MisclassificationEstimate(self, trainIndex, testIndex, classifier = "knn",
  verbose = FALSE, k = 3)
}
\arguments{
\item{classifier}{is a type of the classifier}

\item{k}{is number of neighbours passed to classifier}

\item{countTable}{this data frame for RNA-Seq data which contains one row for indiviual and one column for variable}

\item{classes}{such is vector for our cases for classes,}

\item{trainingproportion}{is the ratio of the training subset from the whole data set}

\item{trainIndex=NULL}{vector providing training sample indices. if NULL, training indices are sampled randomly}
}
\description{
this script to evaluate and assesse the performance of the RNA-Seq classifier by
# Random sampling (random partitioning) estimation of the misclassification rate.
}
\author{
Mustafa ABUELQUMSAN and Jacques van Helden
}
