% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalize_count_table.R
\name{NormalizeSamples}
\alias{NormalizeSamples}
\title{scaling RNA-seq count table}
\usage{
NormalizeSamples(self, method = "quantile", quantile = 0.75, log2 = TRUE,
  epsilon = 0.1)
}
\arguments{
\item{self}{an object belonging to the class DataTableWithClasses
This input object contains a count table, with one column per sample and
one row per gene + a phenoTable + sample class specifications.

The phenoTable of the input object may be filtered out adapted it in case samples would be suppressed because they have a scaling factor of 0.}

\item{classLabels}{the class labels associated each sample of the count table. Should be provided in order to adapt it in case samples would be suppressed because they have a scaling factor of 0.}

\item{method="quantile"}{normalization method. Supported: sum, mean, media, quantile.
Preferred method: quantile = 0.75. Quantiles are more robust to outliers than the mean
or sum. The median is sometimes weak in RNA-seq data, the 75th percentile (quantile 0.75)
seems a good tradeoff between robustness and representativity (taking into account a
representative proportion of the total counts per samples).
Note that median is the same as quantile 0.5.}

\item{quantile=0.75}{quantile used as scaling factor when quantile method is selected.}

\item{log2=TRUE}{apply a log2 transformation}

\item{epsilon=0.1}{value added to all counts before applying the log2 transformation
in order to avoid zero counts.}
}
\value{
the function returns an object of class DataTableWithClasses, where the dataTable
coontains the normalised counts. Note that the normalized count table may have a smaller
number of samples than the input count table because if some samples have a scaling
factor of 0 they will be filtered out. In such case, the phenoTable and classLabels will
be filtered out in the same way to ensure consistency  of the attributes of the
returned DataTableWithClass object.
}
\description{
normalisation of RNA-seq count table.
More precisely this function runs a sample-wise scaling so that all the samples
have the same value for a user-defined scaling parameter.
By default, we use the quantile 0.75 as scaling factor.
}
\examples{

}
\author{
Jacques van Helden and Mustafa AbuElqumsan
}
