% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/count_table_with_train_test_sets.R
\name{print.countTableWithTrainTestSets}
\alias{print.countTableWithTrainTestSets}
\title{print for a countTableWithTrainTestSets object}
\usage{
\method{print}{countTableWithTrainTestSets}(x)
}
