% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/study_case.R
\name{RunDESeq2}
\alias{RunDESeq2}
\title{run DESeq2 to test differential expression on each feature of a data table.}
\usage{
RunDESeq2(self)
}
\arguments{
\item{self}{object belong to StudyCase class}
}
\value{
an object of the same class as the input object
}
\description{
run DESeq2 on an object of class StudyCase to test differential expression on each feature of a data table, and order variables by increasing adjusted p-value.
}
\author{
Mustafa AbuElQumsan and Jacques van Helden
}
