% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/study_case.R
\name{RunViRf}
\alias{RunViRf}
\title{run variable importance by random Forest to on an object of class StudyCase.}
\usage{
RunViRf(self)
}
\arguments{
\item{self}{object belong to StudyCase class.}
}
\value{
an object of the same class as the input object
}
\description{
run variable importance by random Forest to test importance on each feature of a data table.
}
\author{
Mustafa AbuElQumsan and Jacques van Helden
}
