% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/study_case.R
\name{RunViRf.StudyCase}
\alias{RunViRf.StudyCase}
\title{Sort variables by decreasing variable importance as computed by Random Forest}
\usage{
\method{RunViRf}{StudyCase}(self)
}
\arguments{
\item{self}{an object belonging to StudyCase class.}
}
\value{
a clone of the input StudyCase object with an added
DataTableWithTrainTestSets containing the log2-normalized counts data table
where features have been re-ordered by decreasing the importance of the features.
}
\description{
run Random Forest on an object of class StudyCase, and, and order a dataTable by decreasing order of RF variable importance.
}
\author{
Mustafa AbuElQumsan and Jacques van Helden
}
