% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/memory_images.R
\name{saveMemoryImage}
\alias{saveMemoryImage}
\title{Save a memory image in order to be able to reload all results without having to recompute everything}
\usage{
saveMemoryImage(memory.dir = parameters$dir$memoryImages,
  recountID = parameters$recountID, prefix = "RNAseqMVA_image_")
}
\arguments{
\item{memory.dir}{= parameters$dir$memoryImages it is the path the saving the image file.}

\item{recountID}{= parameters$recountID Recount ID number.}

\item{prefix}{= "RNAseqMVA_image_" such prefix in order to saving the file in this prefix.}
}
\description{
if you want to saving time without of made recomputation thoughout loading last computation process.
}
\author{
Jacques van Helden anf Mustafa AbuElQumsan
}
