% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/study_case.R
\name{StudyCase}
\alias{StudyCase}
\title{Create a StudyCase object, load RecountID dataset, and generate datasets.}
\usage{
StudyCase(recountID, parameters)
}
\arguments{
\item{recountID}{A valide ID for an object of the ReCount2 database}

\item{parameters}{recountID-specific parameters specified in a YAML-formatted configuration file}
}
\value{
an object of class StudyCase, that is contain raw data before per-processing step, datasets for testing that are per-processed dataset and their respective parameters.
}
\description{
Create a StudyCase object, load RecountID dataset, and run preprocessing to generate the DataTableWitthClasses objects.
}
\author{
Jacques van Helden and Mustafa AbuElQumsan
}
