% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/countTableWithTrainTestSets.R
\name{summary.countTableWithTrainTestSets}
\alias{summary.countTableWithTrainTestSets}
\title{summary for a countTableWithTrainTestSets object}
\usage{
summary.countTableWithTrainTestSets(x)
}
