% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data_table_with_classes.R
\name{summary.DataTableWithClasses}
\alias{summary.DataTableWithClasses}
\title{print a summary of an object belonging to class DataTableWithClasses}
\usage{
summary.DataTableWithClasses(self)
}
\value{
print the summary of such object by utilizing generic function
}
\description{
just print the summary of the object that is belonge to class DataTableWithClasses
}
\author{
Mustafa AubElQumsan and Jacques van Helden
}
