% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data_table_with_train_test_sets.R
\name{summary.DataTableWithTrainTestSets}
\alias{summary.DataTableWithTrainTestSets}
\title{print a summary of an object belonging to class DataTableWithTrainTestSets}
\usage{
\method{summary}{DataTableWithTrainTestSets}(x)
}
\arguments{
\item{self}{is the targeted object for print its properities.}
}
\value{
print the summary of such object that is belonge to DataTableWithTrainTestSets class by utilizing generic function.
}
\description{
just print the summary of the object that is belonge to class DataTableWithTrainTestSets.
}
\author{
Mustafa AbuELQumsan and Jacques van Helden.
}
