% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/train_test_result.R
\name{TrainTestResult}
\alias{TrainTestResult}
\title{Class to handle the results of Training-Testing evaluation for a given classifier}
\usage{
TrainTestResult(dataset, classifier, iteration, trainPredictedClasses,
  testPredictedClasses)
}
\arguments{
\item{dataset}{the dataset on which the classifier ran}

\item{classifier}{the classifier used to produce this result}

\item{iteration}{is the number of resampling for sake of much more precision to evaluate used classifier.}

\item{trainPredictedClasses}{are the classes that are predicted based on training the classifier with trainIndices.}

\item{testPredicatedClasses}{are the classes that are predicated based on training an classifier with testIndices.}

\item{call}{the function call used to produce this result}
}
\value{
an object of class TrainTestResult containing the following attributes
\itemize {
  \item ID: is that recountID for the test dataset.
  \item classLabels: are labels (classes) for each sample in the targeted dataset.
  \item dataType: are the feature type (preprocessing created data) to analyize the imapct of pre-processing into efficiency of classifer, e.g. (filtered, norm, log2norm, log2norm-Pcs, ....)
  \item parameters: are all general parameter joind with each dataset from methods init_recount_id().
  \item classifier: is the classifier used to produce such results.
  \item iteration: is number of resampling to individuals in order to paly it with classifier.
  \item trainIndices: are the whole indeices for the trainset to be used as the trainset for the classifier ran.
  \item testIndices: are the whole indices for the testset to be paased to classifier as the tesset part.
  \item trainPredictedClasses: all the classes that are predicted based on training the classifier with trainset.
  \item testPredictedClasses: all the classes that are predicted based on training the classifier with testset.
  \item test.contingency: is (testing error) that is a table which contains in raw the actual classes on the testing set and in column predicted classes from classifier on the testing set.wherein that in diagolan is the correct classification and out of diagonal is miscalssifcation error rate for the testing set.
  \item testing.errors: are the class labels that are existed in the actuall classes for the testset and  which are not existed in the predicted classes from the trained classifier.
  \item testing.error.nb: is the sume of the testing.errors.
  \item testing.error.rate: is the training.error.nb divided by whole counts of the testing errors.
  \item train.contingency: is (learning error) the table contain in the raw actual class and in column predicted class and in diagonal is the corect classification and out of diagonal is misclassification errors.
  \item training.errors: are the class labels that are founded in the actuall classes for the trainset and which are not existed in the predicted classes from the trained classifier.
  \item training.error.nb: is the sum of the training.errors
  \item training.error.rate: is the training.error.nb divided by overall counts of the training errors.
  \item trainProportion:  the ratio of the trainset from all the dataset.
  \item trainSize:        is volume of the train size from all the dataset it is computed by multiple number of individuals in the trainProportion magnitude.
  \item testSize:         the test size "remained size" from all dataset after we remove the trainSize from targeted dataset
}
}
\description{
Each an object of this class contains the primary results of a training-testing evaluation of a classifier
}
\author{
Jacques van Helden and Mustafa AbuElQumsan
}
