% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/train_test_result.R
\name{TrainTestResult}
\alias{TrainTestResult}
\title{Class to handle the results of  Training-Testing evaluation for a given classifier}
\usage{
TrainTestResult(dataset, classifier, iteration, trainPredictedClasses,
  testPredictedClasses)
}
\arguments{
\item{dataset}{the dataset on which the classifier ran}

\item{classifier}{the classifier used to produce this result}

\item{call}{the function call used to produce this result}
}
\value{
an object of class TrainResult containing the following attributes
\itemize {
  \item ... MUSTAFA WILL DESCRIBE THE ATTRIBUTES ...
}
}
\author{
Jacques van Helden and Mustafa AbuElQumsan
}
